Adding estimate_quantile_at_value()
This change adds the ability to estimate what the quantile of a
particular value is from a t-digest.

This also exposes the centroids from the t-digest, which might be
a change I'll be able to walk back later on (currently is needed
for custom serialization, but may be possible to use serde).